+++ docs/release_notes.rst
@@ -4,6 +4,11 @@ Release notes
 Change log
 ----------
 
+Changes from 0.10.1 to 0.10.2
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+- Removed upper version limits for dependencies
+
 Changes from 0.10.0 to 0.10.1
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
@@ -67,7 +72,7 @@ Changes from 0.6.2 to 0.7.0
   - AEBS products
   - MLI_SHA_2E
 
-- See :ref:`Available parameters for Swarm data` for details of the collection and measurement names
+- See :ref:`Available parameters for Swarm` for details of the collection and measurement names
 - Added :py:meth:`viresclient.SwarmRequest.available_observatories` to query the AUX_OBS collections to identify IAGA codes available within each collection
 
 Changes from 0.6.1 to 0.6.2